Refactor computed options
